Move the stringify implementation to a separate file
